moving map page components into their own folders
