+++ frontend/components/MapHeader/friendsOptions.js
