+++ frontend/components/MapLegend/Legend.js
